Moved Header components into header folder.
