+++ football-manager/src/lib/Header/searchResultsContainer.svelte
